# desktop/openapi.json
{
  "components": {
    "schemas": {
...
      "AutonomyMode": {
        "enum": [
          "approval",
          "auto-revertible",
          "high-autonomy"
        ],
        "title": "AutonomyMode",
        "type": "string"
      },
      "AutonomyModeRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/AutonomyMode"
          }
        },
        "required": [
          "mode"
        ],
        "title": "AutonomyModeRequest",
        "type": "object"
      },
      "AutonomyModeResponse": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/AutonomyMode"
          }
        },
        "required": [
          "mode"
        ],
        "title": "AutonomyModeResponse",
        "type": "object"
      },
...
      "ChangeOperation": {
        "properties": {
          "action": {
            "enum": [
              "create",
              "update",
              "rename"
            ],
            "title": "Action",
            "type": "string"
          },
          "after": {
            "minLength": 1,
            "title": "After",
            "type": "string"
          },
          "base_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Hash"
          },
          "before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "page_id",
          "path",
          "action",
          "after",
          "rationale",
          "citations"
        ],
        "title": "ChangeOperation",
        "type": "object"
      },
      "ChangeSet": {
        "properties": {
          "agent_runtime": {
            "title": "Agent Runtime",
            "type": "string"
          },
          "base_revision": {
            "title": "Base Revision",
            "type": "string"
          },
          "contradictions": {
            "items": {
              "$ref": "#/components/schemas/Contradiction"
            },
            "title": "Contradictions",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/ChangeOperation"
            },
            "minItems": 1,
            "title": "Operations",
            "type": "array"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Ids",
            "type": "array"
          },
          "status": {
            "default": "pending",
            "enum": [
              "pending",
              "applied",
              "rejected",
              "reverted"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "supersedes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes"
          },
          "validation_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "created_at",
          "source_ids",
          "base_revision",
          "agent_runtime",
          "summary",
          "operations",
          "contradictions"
        ],
        "title": "ChangeSet",
        "type": "object"
      },
      "Citation": {
        "properties": {
          "end_ms": {
            "exclusiveMinimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "start_ms": {
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          }
        },
        "required": [
          "source_id",
          "start_ms",
          "end_ms"
        ],
        "title": "Citation",
        "type": "object"
      },
...
      "Contradiction": {
        "properties": {
          "citations_a": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "minItems": 1,
            "title": "Citations A",
            "type": "array"
          },
          "citations_b": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "minItems": 1,
            "title": "Citations B",
            "type": "array"
          },
          "claim_a": {
            "title": "Claim A",
            "type": "string"
          },
          "claim_b": {
            "title": "Claim B",
            "type": "string"
          },
          "requires_approval": {
            "default": true,
            "title": "Requires Approval",
            "type": "boolean"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic",
          "claim_a",
          "claim_b",
          "citations_a",
          "citations_b"
        ],
        "title": "Contradiction",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "cause_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cause Id"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "component": {
            "title": "Component",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "user_message": {
            "title": "User Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "user_message",
          "retryable",
          "component",
          "operation"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorEnvelope"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "MediaReference": {
        "properties": {
          "asset_path": {
            "title": "Asset Path",
            "type": "string"
          },
          "cached": {
            "title": "Cached",
            "type": "boolean"
          },
          "end_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Ms"
          },
          "kind": {
            "enum": [
              "frame",
              "clip",
              "audio"
            ],
            "title": "Kind",
            "type": "string"
          },
          "playable": {
            "title": "Playable",
            "type": "boolean"
          },
          "rendered_end_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rendered End Ms"
          },
          "rendered_start_ms": {
            "title": "Rendered Start Ms",
            "type": "integer"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "start_ms": {
            "title": "Start Ms",
            "type": "integer"
          }
        },
        "required": [
          "source_id",
          "kind",
          "start_ms",
          "rendered_start_ms",
          "asset_path",
          "cached",
          "playable"
        ],
        "title": "MediaReference",
        "type": "object"
      },
...
      "ValidationIssue": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "operation_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Index"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "severity": {
            "enum": [
              "error",
              "warning"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "severity",
          "message"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
      "ValidationResult": {
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "issues"
        ],
        "title": "ValidationResult",
        "type": "object"
      },
...
      "WikiLintIssue": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "severity": {
            "enum": [
              "error",
              "warning"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "severity",
          "message"
        ],
        "title": "WikiLintIssue",
        "type": "object"
      },
      "WikiLintResponse": {
        "properties": {
          "healthy": {
            "title": "Healthy",
            "type": "boolean"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/WikiLintIssue"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "issues",
          "healthy"
        ],
        "title": "WikiLintResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "vid2note",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/media/asset": {
      "get": {
        "operationId": "media_asset_api_v1_media_asset_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Media Asset",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/clip": {
      "get": {
        "operationId": "media_clip_api_v1_media_clip_get",
        "parameters": [
          {
            "in": "query",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_ms",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Start Ms",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "end_ms",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "End Ms",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "buffer_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 30000,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Buffer Ms"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaReference"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Media Clip",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/frame": {
      "get": {
        "operationId": "media_frame_api_v1_media_frame_get",
        "parameters": [
          {
            "in": "query",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timestamp_ms",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Timestamp Ms",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaReference"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Media Frame",
        "tags": [
          "media"
        ]
      }
    },
...
    "/api/v1/wiki/lint": {
      "get": {
        "operationId": "lint_wiki_api_v1_wiki_lint_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiLintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Lint Wiki",
        "tags": [
          "wiki"
        ]
      }
    },
    "/api/v1/wiki/lint/propose": {
      "post": {
        "operationId": "propose_lint_repairs_api_v1_wiki_lint_propose_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Propose Lint Repairs",
        "tags": [
          "wiki"
        ]
      }
    },
    "/api/v1/wiki/policy": {
      "get": {
        "operationId": "get_wiki_policy_api_v1_wiki_policy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutonomyModeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Wiki Policy",
        "tags": [
          "wiki"
        ]
      },
      "put": {
        "operationId": "update_wiki_policy_api_v1_wiki_policy_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutonomyModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutonomyModeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Update Wiki Policy",
        "tags": [
          "wiki"
        ]
      }
    },
...
  }
}